+++ src/api/consumer/lib/init.js
@@ -1,6 +1,15 @@
 /** @flow */
 import { Consumer } from '../../../consumer';
 
-export default function init(absPath: string, noGit: boolean = false): Promise<Consumer> {
-  return Consumer.create(absPath, noGit).then(consumer => consumer.write());
-}
+export default (async function init(
+  absPath: string = process.cwd(),
+  noGit: boolean = false,
+  reset: boolean = false,
+  resetHard: boolean = false
+): Promise<Consumer> {
+  if (reset || resetHard) {
+    await Consumer.reset(absPath, resetHard, noGit);
+  }
+  const consumer: Consumer = await Consumer.create(absPath, noGit);
+  return consumer.write();
+});